refactor(posts): extract shared PostForm from the create page
The create page's form is the same form the edit dialog needs, so the field
layout and validation move into a reusable PostForm that takes its master
data, default values, submit handler and footer from the caller. The create
page keeps only its own submit behaviour.

create-post-schema.ts becomes post-schema.ts one level up, since it is no
longer owned by the upload route, and PostPosterField moves with it.
